requirements.txt
customtkinter==5.2.2
requests==2.31.0
pillow==10.2.0 